refactor: 사용하지 않아도 될 컴포넌트 내부 index.ts 제거
- 선형님께서 요청하신 사항
