--- src/atoms/Emoji/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Emoji';
