+++ src/atoms/index.ts
@@ -1,10 +1,10 @@
-import Button from './Button';
-import Emoji from './Emoji';
-import Image from './Image';
-import Logo from './Logo';
-import Text from './Text';
-import TextArea from './TextArea';
-import Title from './Title';
+import Button from './Button/Button';
+import Emoji from './Emoji/Emoji';
+import Image from './Image/Image';
+import Logo from './Logo/Logo';
+import Text from './Text/Text';
+import TextArea from './TextArea/TextArea';
+import Title from './Title/Title';
 
 export default {
   Button, Emoji, Image, Logo, Text, TextArea, Title,
